refactor: split docs html member rendering
